Implement a "halting" x86 emulator and test it on a forth "rom"
Bootstrapping on baremetal is not for the faint of heart. For brand new
hardware, it is not even guaranteed that the CPU is fully debugged, let
alone that it will read and write correctly to memory or provide sane
access to any other peripherals. The usual software solution to avoid
all hardware related headaches is to use emulation.

Using an emulator allows one the freedom to not have to deal with actual
hardware, which even in the best of circumstances can be pretty messy
involving supply chains, cold solder joints and etching solutions among
other hassles.

On x86, there are plenty of choices for debugging using an emulator. QEMU,
Bochs and Virtualbox are some of the more well known choices. In this
bootstrap exercise though, I'll choose to roll my own emulator which
implements "just enough emulation" as the bootstrap progresses. This is
a forward looking choice since I plan to eventually port this code over
to other arcane architectures where emulation tools may be a luxury
(or may not even exist). My hope is that this emulator will give me a
better means of debugging during bringup.

Emulating the hardware gives me a certain degree of freedom so on the
flip side, to keep things grounded in reality, I'll choose to use the
NASM assembler to convert x86 assembly to machine code.

There's plenty of choices when it comes to assemblers for x86. The usual
"bundled" alternative for an assembler on x86 is GNU "as" which is pretty
much available everywhere - even in busybox or usually as part of the GNU
GCC bundle. For this exercise though, my choice is to use nasm because I
prefer its cleaner syntax.

For building the whole enchilada, I'll choose to use "make" since it is
also available pretty much everywhere. In the makefiles, my philosphical
take is that it is ok to use a fair bit of duplication to keep things
readable rather than use cryptic make variables which need to be looked
up in the docs every time by most of us ordinary mortals.

The emulator is written in C which should pretty much compile anywhere.

For each commit, I'll end with some metadata to track the changes made
in that commit. Forth related changes in this commit is denoted using
	$FORTH bye
since I'm adding support for the Forth word 'bye'. The assembly code
that matches that is:
	$ASSEMBLY halt
since the label at 'halt' implements this functionality. Emulator
changes, in this case for 0xF4, is denoted as:
	$EMULATE F4
since 0xF4 is the x86 opcode for halting the CPU.
If I'm not changing functionality or adding features or fixing bugs
but addressing cleanup/techdebt, that will be denoted as $TECHDEBT

$FORTH bye
$ASSEMBLY halt
$EMULATE F4