One of the core ideas of Forth is to expose the functionality of the underlying system using what is called a "dictionary" (a namespace, if you want to think of it that way).
Currently the only piece of x86 functionality that is provided by the
emulator is the ability to halt the system. This is provided in assembly
at the label "halt". We map this functionality to the equivalent Forth
word "bye". To provide this mapping, I've introduced a simple Perl
script called "genprims" which takes its input from stdin. The format
of the input must be:
	code{
	forth_name assembly_label
		Multi line assembly code
		which will be passed through
		as is
	}code
where forth_name is the Forth word that will be used to refer to the
functionality provided at assembly_label and the rest of the code is
currently uninterpreted and passed through as is.

The genprims script turns this input into assembler code in prims.s and
also generates a mapping file between Forth and assembler in dict.map
(which can be considered a proxy for the Forth dictionary, for now)