Forth is usable even on "really small" systems. These include systems where the ROM/RAM cannot host all of the dictionary content. This is made possible by a technique called "umblical hosting" which works by hosting the Forth dictionary (which gets larger with the addition of each new "word"/definition) on a "beefy host" (called A in the figure) which is connected over serial (or any other communication channel) to the "wimpy system" under development (called B in the figure below)
        ----------            ----------
        |A:Beefy |            |B:Wimpy |  <- test system
        |  Host  |  <------>  |  system|
        ----------    ^       ----------
            ^         L_ serial (or some other communication channel)
            L_ contains dictionary

It is easy to picture the data exchange over the serial line as an
"infinite Turing tape" usable by the test system.

The "genrom" Perl script enables this type of "umblical hosting". It
uses the dictionary (read from a mapping file on the host, which was
generated by "genprims"). It reads the Forth words that need to be
executed on the test system (which is provided in "rom.4th") from
stdin and generates just the "native" code required to run those words.

Currently, with just one byte of ROM usage, we are in no danger of
filling up the ROM. So we dispense with the serial channel, for now.

$BUGMAKE dependency on genprims

While reviewing the older changes, I realized that the dependency on
the genprims script by prims.s wasn't explicitly called out. So I
fixed that as well as part of the current set of makefile changes.