Implement the inner/"next" interpreter
The crux of the Forth interpreter is usually referred to as "next".

A typical processor continuously runs a "fetch;decode;exec" loop.
So we can now imagine that we have created a "Forth processor" which
fetches the next piece of code to be interpreted and then decodes and
execs it by jumping to the appropriate chunk of native code.

One easy way to visualize this is to think of the x86 CPU as a high
performance interpreter for byte codes. All we have done is add a
layer of indirection to that.

Since the emulator is not yet setup to handle the new opcodes introduced
by these changes, they are currently commented out. As I add functionality
to the emulator to handle these instructions, I'll uncomment them one by
one so that the make is regression free.

$ASSEMBLY inner but currently commented out