Extend the emulator to handle "mov si, immediate"
Although the x86 register SI doesn't provide access to the high
and low bytes within it, I'm looking ahead to the other x86 registers
that will need this access and so I created a "split register structure".

The choice of the x86 SI register for the IP register in Forth is driven
by the fact that lods is the only x86 instruction that allows load with
auto increment and it uses register SI implicitly. The lods instruction
also uses register AX as the implicit destination. So that makes AX the
equivalent of Forth's W register.

$EMULATE BE