Extend the emulator to handle "mov $reg, immediate"
While testing "jmp ax" in the inner interpreter, I found a bug in
the existing assembly code: register AH was not being initialized
but was used in "jmp ax". The "lodsb" prior to the "jmp ax" only
fills in register AL. Fixing that bug required the addition of a
"mov ax, ..." to intialize/restore AH which required opcode 0xB8.
To support that, I needed to extend the emulator. Rather than code
up another special case for register AX just as I had done for SI,
I then decided to make the code generic for all the registers
which required even more changes.

$EMULATE B8 .. BF
$BUGASSEMBLY register AH needs to be initialized before use
$DEBUG struct pointer vs local struct update. I needed to use
	x86register *x=&regs[r];
instead of:
	x86register x=regs[r];
otherwise the updates were made locally instead of the array.