Extend the emulator to handle "jmp $reg"
During the implementation of this change, I ran into a couple of bugs.
Rather than smush all the changes that were needed to fix those bugs
into one single big commit, I split the fix for the pre-existing bugs
into the previous commit. There was a subtle bug in the code that was
being prepared for this commit as well: I used "break" instead of
"return" after changing the instruction pointer. Due to the interactions
between the previous bugs and this bug, I had to spend quite some time
untangling the control flow to figure out why things weren't working
as expected. This bug was noticed only by accident while looking at IP
offsets and the jump byte offsets which were printed out for debugging.

$EMULATE FF (E0 .. E8)
$DEBUG use "return" instead of "break" after changing the control flow