Implement the primitive "key" which reads a character from stdin
On x86, serial port access is trivial: just read in from port 0x3f8
I've taken the easy way out in the emulator and just call a getchar()
and ignore the port (for now)

As part of this change, I also extended the DSL implemented by genprims
to also accept inner{ ... }inner which works exactly like code{ ... }code
except for the implicit addition of a "next" instruction in the assembly
code. For now I've used the macro expansion facility available in nasm
to expand this out into the appropriate jump instruction.

Hiding the implementation of "next" within the DSL gives me the flexibility
to change the implementation of "next" in the future. Currently all that is
needed to implement next is to "jmp inner" but I've chosen to use "jmp bp"
since that allows me to not make any more changes to the emulator.

If branch stalls turn out to be a performance issue it should be fairly
easy to unroll the code at "inner" (ie inline the "lodsb+jmp ax") at the
expense of some code bloat.

$EMULATE EC
$FORTH key
$ASSEMBLY key
$GENPRIMS inner{ ... }inner