Implement the "emit" primitive which outputs a character to the serial port
Just like in the implementation of "key", using the emulator lets me
gloss over the boatload of issues in dealing with real hardware serial
ports by just calling putchar()

The Forth word "emit" pops the stack after it sends the character to the
serial port so in a strict implementation, I'd face the issue of having
to implement the Forth stack before I can implement "emit". But I want to
implement "emit" first since that will help with "printf" style debugging.
So for now, I break this circular dependency by not implementing a stack
but by caching the input value in register bl and using that cached value
during output. So this makes register BX map to Forth's TOS (top of stack)
register.

The emulator needed to be extended to handle the "out"(0xEE) opcode to
generate the output on the serial port. It also needed to be extended to
handle opcode 0x88 (for register to register moves).

The makefile was also modified to verify that the generated output matches
a preset value, so this becomes usable as a regression test as well.

$EMULATE 88 EE
$FORTH emit
$ASSEMBLY emit
$REGISTER TOS:bx IP:si W:ax (also used: dx, bp)