Implement a bootloader that reads in the boot code via the serial port
Now that the serial port has been tested and seen to be working, we can
move on to coding up the bootloader that reads data from the serial port
and stores it in RAM.

As I've done earlier, the protocol is that all the parts of the code that
need emulator support will be commented out. Each of those instructions
will be enabled one at a time as I add support for that instruction to the
emulator. This makes sure that the regression tests pass all the time.

The loader is just a simple loop (using the "rep" x86 prefix) which reads
a byte from the serial port ("insb") and stores it in memory pointed by
the auto incremented register DI.

$ASSEMBLY load