Extend the emulator to handle "cld"
So far, I've played fast and loose with the x86 implementation of lodsb
in the emulator. On real hardware, the x86 direction flag "df" determines
whether the lods and other instructions like it increment or decrement
the pointer to memory. With the addition of the "Clear Direction"("cld")
instruction I've brought the emulator implementation more in line with
the actual hardware implementation.

$EMULATE FC