Extend the emulator to implement "rep insb"
The downside to having a billion transistors at your disposal is the
power bill. The upside, of course, is that you can have loops in a
single instruction. So I've leveraged those billions of transistors
to code up a baremetal bootloader in just a dozen bytes.

Other processors with fewer transistors (which hopefully use fewer watts)
may need a larger number of instructions to implement similar functionality

To simplify the code, I've assumed that the code immediately after the
bootloader is writable and that the control flow continues immediately
at that location. These short cuts may not apply in a real world system.
Also, typical microcontrollers usually have much more space for ROM than
RAM which is the opposite of the scenario that is showcased here.

I've also glossed over messy details such as the handshakes that will be
needed to implement the equivalent of an xmodem/ymodem/zmodem type transfer
since this isn't meant to be much more than a proof of concept.

So there you have it - the start of a Forth implementation which can even
fit in a ROM which is in the (baker's ?) dozens of bytes - assuming there
is a serial port at hand.

$EMULATE 6C F3