Add a means of placing comments in the Forth code in rom.4th
I'm going to break with Forth tradition and use '[' to denote the equivalent
of C/C++ style // comments that go to the end of the line. The choice of '['
quite naturally follows from the fact that it has traditionally been used to
denote the contents of the stack after each word is executed - so at some
level of abstraction, it can be used to mean "the rest of the stack". These
could even be viewed as "structured comments" and you could interpret them
as "Hoare style" proof statements.

So this allows code and comments that look like this:
		[ a b
	dup	[ a b b // make a copy
	drop	[ a b
which means that starting with a stack where 'b' is the element at the top
of the stack (TOS) and 'a' is the element below that (NOS/"next on stack")
dup will duplicate the top element and drop will get rid of it. One is free
to pretend that [ itself represents the rest of the stack. The part after
the // is the usual explanatory comment.

Within a comment, I'll use '<' to denote what the input stream
contains and '>' to denote what the output stream contains so key and emit
could be described as follows:
		[ < c
	key	[ c
	emit	[ > c

This also allows comments to describe the state of the return stack as well
using the convention that ']' represents the "rest of return stack"
For example, the semantics of r> and >r could be described as:
		[ x	|   ]
	>r	[ 	| x ]
	r>	[ x	|   ]

Occasionally I'll also use the following notation when I need to annotate
the state of Forth's IP register (for example to explain lit):
		[ 	| p	|	]
	lit	[ *p	| p+2	|	]

In general, the suggested overall structure of a comment is:
	[SP ... NOS TOS | IP , W | RTOS RNOS ... ]RP <IN ... >OUT ... // ...