Implement dup and drop
Since the x86 stack pointed at by register SP grows to low memory, the area
below the bootloader (0-0x100) is used for the stack. The current memory
map looks like this:

         _____________
0x0:    | Data stack  |
        |_____________|
0x100:  | Bootloader  |
        |_____________|
0x10C:  | Copy of ROM |
        |_____________|

Since dup and drop can be implemented using single byte opcodes, it makes
sense to have that available for use in other places without the overhead
of a branch/call/interpretation. Since they are meant to be private, they
are available as _dup and _drop while the regular Forth names map to the
regular labels dup and drop.

$ASSEMBLY dup drop macros: _dup _drop
$FORTH dup drop