Get rid of some of the clutter accumulating in code.prims
Since much of the code implementing the primitives in code.prims is
boilerplate, it can be compressed down to the one-liner that actually
does the heavy lifting. I've extended the DSL in genprims to introduce
"unary" to describe words that operate only on the TOS. This makes
the pattern of the definitions obvious.

The tradeoff is that genprims now hides most of the implementation details

$GENPRIMS unary
$TECHDEBT