Implement the subtraction(-) operator
For minimality, '-' could just be defined as "neg +", but for performance
the exact same effect is achieved in assembly. The genprims DSL was
expanded to include a "fallthru" directive which achieves that effect.

Since this definition uses the existing neg'ate and add operators, the
emulator did not need any additional functionality

$FORTH -
$ASSEMBLY sub
$GENPRIMS fallthru