Refactor/abstract out the %define and %macro nasm features
Looking to the future and keeping portability in mind, I want to reduce
the dependency of the assembly code on features that are nasm specific.

Rather than use the %define and %macro directives that are available
in nasm, I've added a "define" directive implemented in genprims. For
now, it is just a thin shim layer that generates the equivalent %
directives in nasm. So, all I've added, is a layer of indirection.

I assume this might turn out to be useful if, in the future, this code
is moved to a different architecture where the assembler does not support
this particular syntax (or even the macro feature).

$GENPRIMS define
$TECHDEBT increase portability, maybe