Implement 'swap' which switches the order of the top two elements
On a processor with a better designed instruction set, swap could just
be "xchg tos, [sp]". Of the gazillion instructions that the x86 has, does
it have this one instruction? No. So, of the numerous other choices of
implementation, I've decided to use "nip; xchg; dip" where xchg is another
non-traditional Forth word (like dip) which is meant for internal use only.

Both swap and xchg use the "fallthru" mechanism to avoid branches.

$EMULATE 87 (XX)
$FORTH swap xchg
$ASSEMBLY swap
$GENPRIMS on the mapping line, the part after the label will be passed thru