Implement literals
"literals" are the Forth equivalent of immediate values in machine code

Since the "instruction pointer" for this implementation of Forth is in
register SI, a lodsw along with some register shuffling to preserve the
contents of register AX, is all that is needed.

The regression test output file was modified to handle the change to
the test due to the additional emit. genrom was also modified to handle
single quoted character literals.

I realized that emit and key were coded before the stack was implemented
so their stack signature is incorrect. Rather than clutter this largish
change with even more stuff, I'll fix that in the next commit.

$EMULATE 89 93 AD
$FORTH lit
$ASSEMBLY lit
$GENROM single quote characters