Test and verify that store(!) works
For now, the location in memory that I've chosen to write to (to verify
that the store operator works as expected) is lower in memory than the
expected depth of the stack. This test may inadvertently fail (or worse,
cause silent hard to debug corruption issues) if the stack grows any
deeper (ie beyond what the test currently expects). Just like the test
for fetch in the previous commit, I'll assume that these areas of code
are very unlikely to change in the future, but I'll punt on how exactly
to ensure this. This is clearly tech debt that I'll just have to shoulder
for now and address in a later commit after I have fleshed out the memory
layout issues as well.

$EMULATE 8F07