Implement variables
Conventionally, variables are part of the Forth dictionary but on a small
system with not have enough RAM, there may not be enough free space to
fit in the dictionary along with the data and return stacks. Low end
microcontrollers with just a few bytes of RAM are fairly common in the
bottom end of the price range. Assuming umblical hosting, no space is
required for the dictionary itself, but we still need to have enough RAM
to store sundry variables and the data and return stacks.

At the very minimum, we need a variable that tracks the starting location
of free memory. A trivial "allocator" (if you want to call it that), just
needs to increase that variable to "allocate" memory. The variable that
holds the address of free memory is traditionally called "here" in Forth.

This change adds the "here" variable via the "var" directive which the
genprims script will turn into a placeholder for memory. The genrom script
was also modified to handle references to variables by push'ing their
address on the stack.

Since we now have reserved memory which is safe to write to, I've changed
the test to use "here" as the temporary write area for testing rather
than writing into the data stack.

$GENPRIMS var
$GENROM can handle references to variables
$FORTH here
$ASSEMBLY here
$TECHDEBT no longer need to write to a random location in memory