Implement byte access to memory: cfetch(c@) and cstore(c!)
With the implementation of fetch(@) and store(!), in theory, we have a
"Turing complete" implementation. So from a minimalistic point of view,
this implementation is now done. But from a usability point of view,
I want to have a working REPL, so I'll continue forging ahead.

In the traditional Forth interperter, there is a "state" variable which
is used to signify whether the interpreter is in "compile" state or
"interpret" state. The state variable can be a single byte so to support
byte access, cstore and cfetch have been added (along with the state
variable itself)

While I was there I cleaned up the definitions of fetch and store
to use TOS instead of referring to register bx directly.

There was a bug due to the incomplete implementation of opcode 88, (which
was done very early on), which resulted in a test failure since that opcode
was used by the new code (although the current code itself was correct).
So I had to rework the emulation of opcode 88 to make things sane again
but not before spending a fair bit of time debugging the issue. In
hindsight, I'll chalk that up as tech debt since the emulator by design
is meant to be "just enough emulation" but ideally the parts that the
emulation glossed over should have been flagged as unimplemented rather
than silently returning success.

$FORTH c@ c!
$ASSEMBLY cfetch cstore
$EMULATE 88 (incomplete implementation resulted in bugs, rework required)
$BUG emulation of opcode 88 was incomplete
$TECHDEBT incomplete opcode implementations should call unimpl()