Implement peek(p@) and poke(p!) to read and write I/O mapped regions
The fetch(@) and store(!) primitives cannot be used to access the I/O
mapped regions on x86. So, for completeness, I've added the peek(p@)
and poke(p!) equivalents of @ and ! to read/write the I/O mapped area.

Pretend that this implements a Turing machine with multiple tapes.

In the stack comments, so far, I've used < and > for key and emit.
With additional streams now available, I'll use the convention of
numbered sigils to denote which stream it is.

Rather than save and restore the ah and dx registers which are live,
a jump to restoredx is made which does the equivalent. The "jmp restoredx"
needed an extension to the emulator to handle short (signed) jumps. The
new code is placed near the top so that it remains a short jump.

$EMULATE EB (XX)
$FORTH p@ p!
$ASSEMBLY peek poke