Implement rp@! (and sp@!) and setup the Forth return stack
Like register SI, the only other x86 register that has auto increment
support is register DI. So I'll choose to use that as the return stack
pointer. It is initialized to point to the start of "free memory" and
along with setting up the pointer itself, 100 bytes of space is allocated
for it. For systems with really small RAM, this can be sized smaller.

To initialize the return stack (or to switch it out a different stack),
I've added a word called rp@! (which is a combination of the traditional
Forth words RP@ and RP!)

To maintain symmetry between the data stack and return stack pointers,
I've also added sp@! which does the equivalent for the data stack pointer.

With some careful coding, it might be possible to move the initialization
of the data stack pointer also into rom.4th, but for now, I'll leave it
where it currently is.

$FORTH sp@! rp@!
$ASSEMBLY spxchg rpxchg