Implement the return stack operators tor(>r) and fromr(r>)
One of the distinguishing features of Forth is that it reifies the return
stack so it is under program control rather than being a black box that can
mysteriously cause stack overflows (if the stack happens to be undersized)

Explicit manipulation of the return stack is enabled by tor(>r) (which
moves an element from the data stack to the return stack) and fromr(r>)
which does the reverse, ie move from the return stack to data stack.

Regarding the x86 implementation, "stosw" could be used to store indirectly
(with autoincrement) using register di, but the symmetrical "fetch via di
with autodecrement" does not exist on x86. So the code for tor and fromr
uses explicit increment and decrement of the return stack pointer to keep
the code looking symmetrical.

During testing I again ran into the tech debt issue of incomplete opcode
implementations which silently report success instead of failing.
Hopefully these will be be fewer as things progress.

$FORTH >r r>
$ASSEMBLY tor fromr
$EMULATE 40 .. 4F (and fixup 89 and 8B)
$BUG emulation of opcodes 89 and 8B was incomplete
$TECHDEBT incomplete opcode implementations should call unimpl()