Implement "enter" and "exit" to enable call/return linkage
Although call and return linkage in Forth could be implemented in terms of
>r and r> it is simpler to have separate primitives to do this. Just like
the native instruction to "call" a subroutine, "enter" pushes the current
IP value on the return stack and sets the new value of IP to the "literal"
value in the 2 bytes following it. The equivalent of the native "ret"urn
instruction is "exit" and all it needs to do is pop the previously saved
value from the return stack into IP.

The genrom DSL has been extended to handle def{ and }def which are just
syntactic sugar with the exact same functionality as the conventional
Forth equivalents ':' and ';'

In the makefile the genrom invocation needs to be done twice - once to
generate the actual definitions (and populate the definitions into the
dictionary map) and the second time to generate the rom which may refer
to the definitions created earlier.

For testing, a simple test definition which is called from the rom is used.

$EMULATE AB, 91 .. 97, 8B (35)
$FORTH enter exit
$ASSEMBLY enter exit
$GENROM def{ ... }def