Implement "call" to enable runtime call/return linkage
"call" takes an address on the data stack and runs through the execution
starting at that address.

Like "enter" and "exit", "call" could be implemented in terms of other
primitives but is provided for performance (maybe?)

For testing, the test case is a tiny bit convoluted since that is the easy
means that I can see of getting an address to "call". If you were paying
attention, yes, that was a (semi-)coroutine. Full coroutines need to be
deferred until returns stack switching is in place, but for now this is
more than sufficient to code up "generators" or for two routines to switch
control back and forth by using the sequence shown in the test

Testing uncovered a bug in the emulation of EB(jump, for positive offsets)
(due to the forward "jmp" in "call") which is now fixed.

$FORTH call
$ASSEMBLY call
$BUG emulation of EB for positive offsets was incorrect (off by one)