Implement primitives to enable control flow branching
Forth traditionally has two control flow primitives, an unconditional
"branch" and a conditional "branch0". Since most processors support various
flavors of jump including zero(jz)/non-zero(jnz) jumps, it is useful to
expose those, as is, at the upper layer as well. So, in this implementation
I'll provide one more in addition to the two traditional branch operators:
"jnz" which jumps when the TOS is non-zero.

This is in addition to the usual ones viz "jmp" which is identical to
"br"/"branch" and "jz" which is identical to "br0"/"branch0"

Rather than clutter up this commit with the test verification changes,
I'll defer those to a later commit.

$FORTH jmp jz jnz
$ASSEMBLY branch branchz branchnz