Implement looping control flow constructs
Forth (and Lisp, with macros) are among the tiny set of languages where
the control flow operators (or, many of what would be considered "reserved
words", in a run of the mill language) are not necessarily built in to the
language but can be "bolted on" later. Almost like a DSL (domain specific
language) implemented in itself. Taking the flexibility that a language
such as this allows to its logical extreme, especially with Forth, can
result in problems of interoperability that ANS Forth tried to address
with its standarization effort. Talking about standards, here's ... N+1

Forth has traditionally used BEGIN ... REPEAT for unconditional loops
and BEGIN ... WHILE ... REPEAT for "while true" type conditional loops
and BEGIN ... UNTIL for "until true" type conditional loops

Since I'd prefer to "syntax sugar" my code a little differently, I've
chosen to use a different set of "keywords". Unconditional loops are:
	loop{ ... }loop
Conditional "while true" loops are coded as:
	loop{ ... }while{ ... }loop
and finally, conditional "until true" loops are coded as:
	loop{ ... }until{ ... }loop
mostly because I prefer the C-style bracing and indentation, although
now the braces need to get attached to the keywords due to the Forth
rule that spaces are the only delimiters between words.

I've used these new looping constructs to code "getc" which loops until the
ready bit on the serial port is set and then reads a character from the
serial port. The "echon" test word showcases the use of the while loop
construct by reading and echoing back n characters from the serial port.

Testing these words is deferred to a later commit since the current changes
are meant to verify that the changes to genrom are working as expected.
All that I've confirmed for now is that the code generated by genrom for
the loops looks mostly correct (by looking at the generated assembly code)

$FORTH getc echon
$GENROM loop{ }while{ }until{ }loop