Fix bug in hexadecimal handling in genrom (and wrong constant in rom.4th)
While testing the changes that were made to genrom, specifically, why
the 0x3f8 port was not being stored correctly, I noticed that there
was a bug in the hexadecimal handling (which has been in the code for
a long while) which didn't get caught due to another bug in the testsuite
in rom.4th which was using 100 instead of 0x100. Sigh!

That was a couple hours worth of debugging for a stupid bug.

$BUG genrom wasn't handling hexadecimals correctly
$BUG rom.4th did not use the right offset to the start of rom code