Verify functionality of loop{ ... }until{ ... }loop in getc
Although the getc code which uses the "until true" loop structure was added
in a previous commit, it wasn't used/tested until now. So when it was
enabled for testing, a whole raft of bugs appeared to crawl out of nowhere.

First I went through a couple of bugs in the emulation of "cbw" which
was because its implementation was trickier than I expected it to be.

Then there was some functionality that needed to be added to really emulate
"p@" since the initial implementation was calling getchar() in all cases.

After all of that was ironed out, I ran into a strange memory corruption
issue which needed me to roll out the heavy guns for debugging. Since I'm
using an emulator, the software equivalent of using a logic analyzer with
a conditional breakpoint is fairly easy to achieve. All of the debug
code I added is under a DEBUG #define flag and it helped me to figure out
the root cause which was a bug in the emulation of "fetch" which uses the
x86 instruction "mov bx, [bx]". The emulator did not take the aliasing into
consideration resulting in the memory corruption bug that I encountered.

While I was there, I also added code to print out the contents of the data
stack after each instruction which also turned out to be pretty helpful in
debugging this issue.

Next step is to verify the "while true" loop structure.

$EMULATE 72
$TECHDEBT emulation of negate needed to be extended to handle carry flag
$BUG emulation of in al, dw was incomplete and needed to be fleshed out
$BUG emulation of cbw, went through a couple of iterations to get correct
$BUG emulation of mov bx, [bx] caused a corruption due to aliasing issues