Verify functionality of loop{ ... }while{ ... }loop in echon
Testing the "while true" functionality used in echon turned up yet another
"off by one" bug in the emulation code for cbw. Once that was fixed, things
worked as expected.

The emulator needed to be extended yet again to handle jnc(0x73). This was
in addition to the emulation for opcode jc(0x72) which was added as part of
the previous commit.

$EMULATE 73
$BUG Off by one error in the offset calculation in the emulation of cbw