Implement conditional control flow constructs
Just like with the "braced" control flow looping keywords, I've sprinkled
some "brace syntax sugar" on the conditional control flow constructs of
traditional Forth. So instead of IF ... THEN I'll be using if{ ... }if

So far, without the availability of condition checks, all of the testing
has relied on generating "positive" output and comparing that to a control
file. Now that conditionals are available, asserts can be coded with the
opposite behaviour ie no additional output counts as no new regressions.

The conditional control flow functionality is tested using this approach.

$GENROM if{ }if