Reimplement "over", this time as native code (and do some cleanup)
Although "over" can be a defined word, I think it might be performance
critical enough that it belongs to the set of words that are defined
natively. So I've reimplemented it in x86 assembly instead of the
earlier definition as "1 pick" that was used in the previous commit.

While re-implementing "over", I found that the code for "fetch" would need
to be duplicated. So, "fetch" has also been turned into a "define macro"
to avoid the need for code duplication.

While I was there, I found some direct references to the underlying x86
registers which were already define'd to the Forth equivalents so I
cleaned up those references as well.

$FORTH over
$ASSEMBLY over