Implement counted loops using for{ ... }for
Forth traditionally provides counted loops using DO ... LOOP and
DO ... +LOOP

Just as I did with the regular loop structures, I've changed the syntax
just a wee bit so as to be able to use the braced versions: for{ ... }for
instead. It takes 2 parameters from the stack viz "start" and "end" and
loops in the range [start,end)
For now, if start > end, the loop can go on forever: [start,end+2^N)

Also defined in this commit is "i" which returns the current index.

While testing these changes, I ran into an existing bug in genrom which
caused duplicated labels to be generated. This was because genrom wasn't
passing the "used label" information between invocations. This is also now
fixed so that the last unused label is passed to the subsequent invocation.

$FORTH i
$GENROM for{ }for
$BUG genrom needs to pass last unused label to later invocations