Implement Henry Massalin's "superoptimized" signum(~) operator
Given elements a and b on the stack (where a is the tos), the signum
aka '~' operator will return 1 if (a>b), 0 if (a==b) and -1 if (a<b)

This branch free code is based on Henry Massalin's superoptimizer technique

While testing this set of changes, I found that printf on a bitfield gives
non-intuitive results. So I got rid of the bitfield and replaced it with a
plain old int which has known semantics.

$FORTH ~
$ASSEMBLY signum
$EMULATE 11 19 29
$GENROM support -ve decimals and hexadecimals