Implement the comparison operators iszero(0=) gt(>) and lt(<)
The iszero(0=) comparison operator returns (tos ? 0 : 1) and is yet another
superoptimized code version (like the signum comparison operator).

The lt(<) and gt(>) comparison operators are defined in terms of the ge(>=)
and le(<=) along with the iszero(0=) comparison operator which keeps the
definitions simple at the cost of some performance.

$FORTH 0= > <
$ASSEMBLY iszero