Implement the "rot" operator which bubbles up the third element to tos
"rot" it another one of the stack manipulation operators which is used
often enough that I assume the performance matters enough to make it a
native code word (although it is trivial enough to implement it in an
architecture neutral way as ">r swap r> swap")

With the addition of the "rot" operator, the 256 byte area reserved for
the native code words is now, for all practical purposes, full. Since
signum is the largest of the native code words, I've placed it at the end
so that its definition can dangle off the end of the 256 byte area, to get
maximal packing density. Since signum is 11 bytes long, I could add another
6 bytes worth of native code before I need to think about moving stuff out
of the 256 byte area. If push comes to shove, I could move out key and emit
since they are not performance sensitive.

$FORTH rot
$ASSEMBLY rot