Implement "alloc" to allocate memory
This code is identical to the code used earlier for the initial allocation
of the return stack. Note that alloc cannot be used there despite the fact
that it is available now because "alloc" is a defined word and the return
stack needs to be in place prior to any calls to defined words. I've chosen
to use a little code duplication to break this circular dependency.

I've also chosen to call the memory allocation operator "alloc" instead of
the traditional Forth word ALLOT since that seems "more correct".

After alloc'ating memory (for multiple return stacks), it is possible to
have multiple threads of execution (and/or "full" coroutines) by switching
between the return stacks using "rp@!"

Independent stacks for each of those execution contexts is also possible by
switching between the data stacks using "sp@!", so it is clear that we now
have a runtime powerful enough to support multi-threaded code, all of which
fits in just ~300 bytes (excluding the test code, of course).

$FORTH alloc