Add instrumentation in the emulator to measure the data/return stack usage
Resource usage of an implementation can be measured along a couple of axes:
1. Flash/ROM usage (bytes, or higher units)
2. RAM usage (bytes, or higher units)
3. CPU cycle (instructions, memory cycles etc)

The Flash/ROM usage requirement of this implementation is fairly easy to
obtain - just look at the tail end offsets in the assembly listing
(specifically, looking at x86/forth.lst, generated by the nasm assembler)

Currently the last byte of "code" is located at offset 0x355 so the code
size/ROM usage clocks in at ~850 bytes (including the regression tests).
I'll estimate that, without the tests, ROM usage would only be ~300 bytes.

If the 12 byte bootloader on x86 was used to bootstrap using a serial port,
this requirement would shift even lower to just 12 bytes of ROM and some
additional RAM or flash to store the bootstrap content that was loaded in
from the "umblical host" or even a SPI Flash/ROM part.

Estimating the RAM requirements is a little harder. Traditionally Forth
uses the RAM for:
1. Data stack(s)
2. Return stack(s)
3. Dictionary
4. "other" book keeping

In the current implementation since I'm using "umblical hosting", we
don't need to worry about the RAM used by the dictionary itself but there
are variables (such as "here") in my implementation, which for now can be
counted manually to arrive at a requirement of 3 bytes:
2 bytes for "here" and 1 byte for "state"

Since it is useful to know the RAM requirements for the stacks as well,
I've added some instrumention in the emulator to be able to easily measure
the data and return stack usage so as to "right size" the stack allocations
From the output generated while running this code, the current RAM usage
for the data stack appears to be 14 bytes (this includes all of the current
regression tests) while the return stack uses just 6 bytes (out of the
100 bytes allocated to it). So the overall RAM usage currently is:
3+14+6=23 bytes.

For the CPU cycle estimates, most decent emulators provide that information
but since this is a "just enough emulation" effort, I will not attempt to
gather that information, for now.